Added GPU particle advance and test routines
Implemented the GPU padvnc and two test routines to verify that the GPU routine is giving the correct answer. This version looks good, the one issue is that the time step control may be implemented incorrectly.

Still need to implement collisions and I think that the reinjection routines should be ported to the gpu.