Finished implementing GPU particle advance and fully functional serial GPU runs.
Scpetic3D can be run on a GPU instead of a CPU by using the command line option --gpurun

MPI version of sceptic3Dgpu is mostly working, however, sceptic3D needs to be modified to allow for each MPI thread to handle a different number of particles in order to get the maximum benefit.

Currently working on moving some of the arrays to texture memory.